Refactor code for cleaner inheritance of props for snapshots
